+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ### Breaking changes
 
+* RealmChangeListener provides the changed object/Realm/collection as well (#1594).
 * All JSON methods on Realm now only wraps JSONException in RealmException. All other Exceptions are thrown as they are.
 * Marked all methods on `RealmObject` and all public classes final (#1594).
 * Removed `BaseRealm` from the public API.
 * Removed `HandlerController` from the public API.
-* Removed constructor of `RealmAsyncTask` from the public API8 (#1594).
+* Removed constructor of `RealmAsyncTask` from the public API (#1594).
+* `RealmBaseAdapter` has been moved to its own GitHub repository: https://github.com/realm/realm-android-adapters
+  See https://github.com/realm/realm-android-adapters/README.md for further info on how to include it.
 
 ### Deprecated
 
@@ -49,6 +52,7 @@
 * `RealmResults.sort()` and `RealmList.sort()` now return the sorted result instead of sorting in-place.
 * `RealmList.first()` and `RealmList.last()` now throw `ArrayIndexOutOfBoundsException` if `RealmList` is empty.
 * Removed deprecated method `Realm.getTable()` from public API.
+* `Realm.refresh()` and `DynamicRealm.refresh()` on a Looper no longer have any effect. `RealmObject` and `RealmResults` are always updated on the next event loop.
 
 ### Deprecated